Resolve merge conflict

+++ Source/dooble_cookies_window.cc
@@ -27,6 +27,7 @@
 
 #include <QDir>
 #include <QKeyEvent>
+#include <QMessageBox>
 #include <QSqlQuery>
 #include <QStatusBar>
 #include <QWebEngineCookieStore>
@@ -41,13 +42,29 @@
 dooble_cookies_window::dooble_cookies_window(bool is_private, QWidget *parent):
   QMainWindow(parent)
 {
+  m_ui.setupUi(this);
+  m_collapse = new QToolButton(m_ui.tree);
+
+  auto font(m_collapse->font());
+
+  font.setStyleHint(QFont::Courier);
+  m_collapse->resize(25, 25);
+  m_collapse->setCheckable(true);
+  m_collapse->setFont(font);
+  m_collapse->setStyleSheet("QToolButton {border: none;}"
+			    "QToolButton::menu-button {border: none;}");
+  m_collapse->setText(tr("+"));
+  m_collapse->setToolTip(tr("Collapse / Expand"));
   m_domain_filter_timer.setInterval(750);
   m_domain_filter_timer.setSingleShot(true);
   m_is_private = is_private;
   m_purge_domains_timer.setInterval(30000);
-  m_ui.setupUi(this);
+  m_ui.block_subdomains->setChecked
+    (dooble_settings::setting("cookies_block_subdomains").toBool());
+  m_ui.blocked->setCheckState(Qt::Checked);
   m_ui.domain->setText("");
   m_ui.expiration_date->setText("");
+  m_ui.favorite->setCheckState(Qt::PartiallyChecked);
   m_ui.name->setText("");
   m_ui.path->setText("");
   m_ui.periodically_purge->setChecked
@@ -58,6 +75,9 @@ dooble_cookies_window::dooble_cookies_window(bool is_private, QWidget *parent):
 
   m_ui.toggle_shown->setProperty("state", true);
   m_ui.tool_bar->addWidget(m_ui.periodically_purge);
+  m_ui.tree->header()->setDefaultAlignment(Qt::AlignCenter);
+  m_ui.tree->header()->setMinimumHeight(30);
+  m_ui.tree->setMinimumWidth(200);
   m_ui.tree->sortItems(0, Qt::AscendingOrder);
   m_ui.value->setText("");
 
@@ -73,6 +93,7 @@ dooble_cookies_window::dooble_cookies_window(bool is_private, QWidget *parent):
     (tr("<html>Private cookies exist within "
 	"the scope of this window's parent Dooble window. Neither "
 	"window geometry nor window state will be retained.</html>"));
+  m_collapse->move(5, (m_ui.tree->header()->size().height() - 25) / 2 + 2);
   statusBar()->addPermanentWidget(label);
 
   if(dooble_settings::setting("denote_private_widgets").toBool())
@@ -96,6 +117,22 @@ dooble_cookies_window::dooble_cookies_window(bool is_private, QWidget *parent):
 	  SIGNAL(applied(void)),
 	  this,
 	  SLOT(slot_settings_applied(void)));
+  connect(m_collapse,
+	  SIGNAL(clicked(void)),
+	  this,
+	  SLOT(slot_collapse_all(void)));
+  connect(m_ui.add,
+	  SIGNAL(clicked(void)),
+	  this,
+	  SLOT(slot_add_blocked_domain(void)));
+  connect(m_ui.block_domain,
+	  SIGNAL(returnPressed(void)),
+	  this,
+	  SLOT(slot_add_blocked_domain(void)));
+  connect(m_ui.block_subdomains,
+	  SIGNAL(toggled(bool)),
+	  this,
+	  SLOT(slot_block_subdomains(bool)));
   connect(m_ui.delete_selected,
 	  SIGNAL(clicked(void)),
 	  this,
@@ -137,6 +174,55 @@ dooble_cookies_window::dooble_cookies_window(bool is_private, QWidget *parent):
   setContextMenuPolicy(Qt::NoContextMenu);
 }
 
+bool dooble_cookies_window::is_domain_blocked(const QUrl &url) const
+{
+  if(url.isEmpty() || !url.isValid())
+    return false;
+
+  if(m_ui.block_subdomains->isChecked())
+    {
+      QStringList domains;
+      auto host(url.host());
+
+      if(!host.startsWith("."))
+	domains << "." + host;
+
+      domains << host;
+
+      while(host.contains('.'))
+	{
+	  auto index = host.indexOf('.');
+
+	  host = host.mid(index + 1);
+
+	  if(!host.isEmpty())
+	    domains << "." + host << host;
+	}
+
+      foreach(const auto &domain, domains)
+	{
+	  auto item = m_top_level_items.value(domain);
+
+	  if(item && item->checkState(0) == Qt::Checked)
+	    return true;
+	}
+    }
+  else
+    {
+      auto item = m_top_level_items.value("." + url.host());
+
+      if(item && item->checkState(0) == Qt::Checked)
+	return true;
+
+      item = m_top_level_items.value(url.host());
+
+      if(item && item->checkState(0) == Qt::Checked)
+	return true;
+    }
+
+  return false;
+}
+
 void dooble_cookies_window::closeEvent(QCloseEvent *event)
 {
   if(!m_is_private)
@@ -149,6 +235,9 @@ void dooble_cookies_window::closeEvent(QCloseEvent *event)
 void dooble_cookies_window::delete_top_level_items
 (const QList<QTreeWidgetItem *> &list)
 {
+  if(list.isEmpty())
+    return;
+
   QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
 
   if(m_cookie_store && m_cookies)
@@ -160,7 +249,7 @@ void dooble_cookies_window::delete_top_level_items
   QList<QNetworkCookie> cookies;
   QStringList domains;
 
-  for(auto item : list)
+  foreach(auto item, list)
     {
       if(!item)
 	continue;
@@ -235,6 +324,12 @@ void dooble_cookies_window::keyPressEvent(QKeyEvent *event)
 void dooble_cookies_window::resizeEvent(QResizeEvent *event)
 {
   QMainWindow::resizeEvent(event);
+  m_ui.value->setText
+    (m_ui.value->fontMetrics().
+     elidedText(m_ui.value->property("value").toString(),
+		Qt::ElideRight,
+		m_ui.value->width()));
+  m_ui.value->setCursorPosition(0);
   save_settings();
 }
 
@@ -251,12 +346,13 @@ void dooble_cookies_window::save_settings(void)
     ("dooble_cookies_window_state", saveState().toBase64());
 }
 
-void dooble_cookies_window::setCookieStore(QWebEngineCookieStore *cookie_store)
+void dooble_cookies_window::set_cookie_store
+(QWebEngineCookieStore *cookie_store)
 {
   m_cookie_store = cookie_store;
 }
 
-void dooble_cookies_window::setCookies(dooble_cookies *cookies)
+void dooble_cookies_window::set_cookies(dooble_cookies *cookies)
 {
   m_cookies = cookies;
 }
@@ -295,6 +391,63 @@ void dooble_cookies_window::show_normal(QWidget *parent)
   QMainWindow::showNormal();
 }
 
+void dooble_cookies_window::slot_add_blocked_domain(void)
+{
+  auto domain(m_ui.block_domain->text().trimmed());
+
+  if(domain.length() <= 1 || m_top_level_items.value(domain))
+    return;
+
+  auto item = new QTreeWidgetItem(m_ui.tree, QStringList() << domain);
+  auto text(m_ui.domain_filter->text().toLower().trimmed());
+
+  item->setCheckState(0, Qt::Checked);
+  item->setFlags(Qt::ItemIsEnabled |
+		 Qt::ItemIsSelectable |
+		 Qt::ItemIsUserCheckable |
+		 Qt::ItemIsUserTristate);
+
+  if(!text.isEmpty())
+    item->setHidden(!domain.contains(text));
+
+  m_top_level_items[domain] = item;
+  disconnect(m_ui.tree,
+	     SIGNAL(itemChanged(QTreeWidgetItem *, int)),
+	     this,
+	     SLOT(slot_item_changed(QTreeWidgetItem *, int)));
+  m_ui.tree->addTopLevelItem(item);
+  m_ui.tree->sortItems
+    (m_ui.tree->sortColumn(), m_ui.tree->header()->sortIndicatorOrder());
+  m_ui.tree->resizeColumnToContents(0);
+  connect(m_ui.tree,
+	  SIGNAL(itemChanged(QTreeWidgetItem *, int)),
+	  this,
+	  SLOT(slot_item_changed(QTreeWidgetItem *, int)));
+}
+
+void dooble_cookies_window::slot_block_subdomains(bool state)
+{
+  dooble_settings::set_setting("cookies_block_subdomains", state);
+}
+
+void dooble_cookies_window::slot_collapse_all(void)
+{
+  QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
+
+  if(m_collapse->isChecked())
+    {
+      m_collapse->setText(tr("-"));
+      m_ui.tree->expandAll();
+    }
+  else
+    {
+      m_collapse->setText(tr("+"));
+      m_ui.tree->collapseAll();
+    }
+
+  QApplication::restoreOverrideCursor();
+}
+
 void dooble_cookies_window::slot_cookie_removed(const QNetworkCookie &cookie)
 {
   auto hash(m_child_items.value(cookie.domain()));
@@ -303,7 +456,7 @@ void dooble_cookies_window::slot_cookie_removed(const QNetworkCookie &cookie)
     {
       auto item = m_top_level_items.value(cookie.domain());
 
-      if(item && item->checkState(0) != Qt::Checked)
+      if(item && item->checkState(0) == Qt::Unchecked)
 	{
 	  m_top_level_items.remove(cookie.domain());
 	  delete m_ui.tree->takeTopLevelItem
@@ -325,20 +478,20 @@ void dooble_cookies_window::slot_cookie_removed(const QNetworkCookie &cookie)
       m_child_items.remove(cookie.domain());
       item = m_top_level_items.value(cookie.domain());
 
-      if(item && item->checkState(0) != Qt::Checked)
+      if(item && item->checkState(0) == Qt::Unchecked)
 	{
 	  m_top_level_items.remove(cookie.domain());
 	  delete m_ui.tree->takeTopLevelItem
 	    (m_ui.tree->indexOfTopLevelItem(item));
 	}
     }
-  else
+  else if(!cookie.domain().trimmed().isEmpty())
     m_child_items[cookie.domain()] = hash;
 }
 
 void dooble_cookies_window::slot_cookies_added
 (const QList<QNetworkCookie> &cookies,
- const QList<bool> &is_favorites)
+ const QList<int> &is_blocked_or_favorite)
 {
   disconnect(m_ui.tree,
 	     SIGNAL(itemChanged(QTreeWidgetItem *, int)),
@@ -347,8 +500,10 @@ void dooble_cookies_window::slot_cookies_added
 
   for(int i = 0; i < cookies.size(); i++)
     {
-      auto is_favorite = is_favorites.value(i);
       const auto &cookie(cookies.at(i));
+
+      if(cookie.domain().trimmed().isEmpty())
+	continue;
 
       if(!m_top_level_items.contains(cookie.domain()))
 	{
@@ -356,15 +511,20 @@ void dooble_cookies_window::slot_cookies_added
 	    (m_ui.tree, QStringList() << cookie.domain());
 	  auto text(m_ui.domain_filter->text().toLower().trimmed());
 
-	  if(is_favorite)
+	  if(is_blocked_or_favorite.at(i) ==
+	     dooble_cookies::BlockedOrFavorite::BLOCKED)
 	    item->setCheckState(0, Qt::Checked);
+	  else if(is_blocked_or_favorite.at(i) ==
+		  dooble_cookies::BlockedOrFavorite::FAVORITE)
+	    item->setCheckState(0, Qt::PartiallyChecked);
 	  else
 	    item->setCheckState(0, Qt::Unchecked);
 
 	  item->setData(1, Qt::UserRole, cookie.toRawForm());
 	  item->setFlags(Qt::ItemIsEnabled |
 			 Qt::ItemIsSelectable |
-			 Qt::ItemIsUserCheckable);
+			 Qt::ItemIsUserCheckable |
+			 Qt::ItemIsUserTristate);
 
 	  if(!text.isEmpty())
 	    item->setHidden(!cookie.domain().contains(text));
@@ -415,7 +575,27 @@ void dooble_cookies_window::slot_cookies_cleared(void)
 
 void dooble_cookies_window::slot_delete_selected(void)
 {
+  if(!m_ui.tree->selectionModel()->hasSelection())
+    return;
+
+  QMessageBox mb(this);
+
+  mb.setIcon(QMessageBox::Question);
+  mb.setStandardButtons(QMessageBox::No | QMessageBox::Yes);
+  mb.setText(tr("Delete selected?"));
+  mb.setWindowIcon(windowIcon());
+  mb.setWindowModality(Qt::ApplicationModal);
+  mb.setWindowTitle(tr("Dooble: Confirmation"));
+
+  if(mb.exec() != QMessageBox::Yes)
+    {
+      QApplication::processEvents();
+      return;
+    }
+
   QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
+
+  auto list(m_ui.tree->selectedItems());
 
   if(m_cookie_store && m_cookies)
     disconnect(m_cookie_store,
@@ -425,9 +605,8 @@ void dooble_cookies_window::slot_delete_selected(void)
 
   QList<QNetworkCookie> cookies;
   QStringList domains;
-  auto list(m_ui.tree->selectedItems());
 
-  for(auto item : list)
+  foreach(auto item, list)
     {
       if(!item)
 	continue;
@@ -521,22 +700,100 @@ void dooble_cookies_window::slot_delete_shown(void)
 {
   QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
 
+  auto found = false;
+
+  for(int i = 0; i < m_ui.tree->topLevelItemCount(); i++)
+    {
+      auto item = m_ui.tree->topLevelItem(i);
+
+      if(item && !item->isHidden())
+	{
+	  found = true;
+	  break;
+	}
+    }
+
+  if(!found)
+    {
+      QApplication::restoreOverrideCursor();
+      return;
+    }
+  else
+    QApplication::restoreOverrideCursor();
+
+  QMessageBox mb(this);
+
+  mb.setIcon(QMessageBox::Question);
+  mb.setStandardButtons(QMessageBox::No | QMessageBox::Yes);
+  mb.setText(tr("Delete shown?"));
+  mb.setWindowIcon(windowIcon());
+  mb.setWindowModality(Qt::ApplicationModal);
+  mb.setWindowTitle(tr("Dooble: Confirmation"));
+
+  if(mb.exec() != QMessageBox::Yes)
+    {
+      QApplication::processEvents();
+      return;
+    }
+
+  QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
+
   QList<QTreeWidgetItem *> list;
 
   for(int i = 0; i < m_ui.tree->topLevelItemCount(); i++)
     {
       auto item = m_ui.tree->topLevelItem(i);
 
-      if(!(!item || item->isHidden()))
+      if(item && !item->isHidden())
 	list << item;
     }
 
-  delete_top_level_items(list);
+  if(!list.isEmpty())
+    delete_top_level_items(list);
+
   QApplication::restoreOverrideCursor();
 }
 
 void dooble_cookies_window::slot_delete_unchecked(void)
 {
+  QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
+
+  auto found = false;
+
+  for(int i = 0; i < m_ui.tree->topLevelItemCount(); i++)
+    {
+      auto item = m_ui.tree->topLevelItem(i);
+
+      if(item && item->checkState(0) == Qt::Unchecked && !item->isHidden())
+	{
+	  found = true;
+	  break;
+	}
+    }
+
+  if(!found)
+    {
+      QApplication::restoreOverrideCursor();
+      return;
+    }
+  else
+    QApplication::restoreOverrideCursor();
+
+  QMessageBox mb(this);
+
+  mb.setIcon(QMessageBox::Question);
+  mb.setStandardButtons(QMessageBox::No | QMessageBox::Yes);
+  mb.setText(tr("Delete unchecked?"));
+  mb.setWindowIcon(windowIcon());
+  mb.setWindowModality(Qt::ApplicationModal);
+  mb.setWindowTitle(tr("Dooble: Confirmation"));
+
+  if(mb.exec() != QMessageBox::Yes)
+    {
+      QApplication::processEvents();
+      return;
+    }
+
   QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
 
   QList<QTreeWidgetItem *> list;
@@ -549,7 +806,9 @@ void dooble_cookies_window::slot_delete_unchecked(void)
 	list << item;
     }
 
-  delete_top_level_items(list);
+  if(!list.isEmpty())
+    delete_top_level_items(list);
+
   QApplication::restoreOverrideCursor();
 }
 
@@ -622,10 +881,17 @@ void dooble_cookies_window::slot_item_changed(QTreeWidgetItem *item, int column)
 
 	query.addBindValue
 	  (dooble::s_cryptography->hmac(item->text(0)).toBase64());
-	query.addBindValue
-	  (dooble::s_cryptography->
-	   hmac(item->checkState(0) == Qt::Checked ?
-		QByteArray("true") : QByteArray("false")).toBase64());
+
+	if(item->checkState(0) == Qt::Checked)
+	  query.addBindValue
+	    (dooble::s_cryptography->hmac(QByteArray("blocked")).toBase64());
+	else if(item->checkState(0) == Qt::PartiallyChecked)
+	  query.addBindValue
+	    (dooble::s_cryptography->hmac(QByteArray("favorite")).toBase64());
+	else
+	  query.addBindValue
+	    (dooble::s_cryptography->hmac(QByteArray("xyz")).toBase64());
+
 	query.exec();
       }
 
@@ -644,8 +910,12 @@ void dooble_cookies_window::slot_item_selection_changed(void)
     {
       m_ui.domain->setText("");
       m_ui.expiration_date->setText("");
+      m_ui.http_only->setChecked(false);
       m_ui.name->setText("");
       m_ui.path->setText("");
+      m_ui.secure->setChecked(false);
+      m_ui.session->setChecked(false);
+      m_ui.value->setProperty("value", "");
       m_ui.value->setText("");
       return;
     }
@@ -660,8 +930,12 @@ void dooble_cookies_window::slot_item_selection_changed(void)
       m_ui.domain->setText(item->text(0));
       m_ui.domain->setCursorPosition(0);
       m_ui.expiration_date->setText("");
+      m_ui.http_only->setChecked(false);
       m_ui.name->setText("");
       m_ui.path->setText("");
+      m_ui.secure->setChecked(false);
+      m_ui.session->setChecked(false);
+      m_ui.value->setProperty("value", "");
       m_ui.value->setText("");
       return;
     }
@@ -676,6 +950,7 @@ void dooble_cookies_window::slot_item_selection_changed(void)
       m_ui.path->setText("");
       m_ui.secure->setChecked(false);
       m_ui.session->setChecked(false);
+      m_ui.value->setProperty("value", "");
       m_ui.value->setText("");
     }
   else
@@ -691,7 +966,12 @@ void dooble_cookies_window::slot_item_selection_changed(void)
       m_ui.path->setCursorPosition(0);
       m_ui.secure->setChecked(cookie.at(0).isSecure());
       m_ui.session->setChecked(cookie.at(0).isSessionCookie());
-      m_ui.value->setText(cookie.at(0).value());
+      m_ui.value->setProperty("value", cookie.at(0).value());
+      m_ui.value->setText
+	(m_ui.value->fontMetrics().elidedText(cookie.at(0).value(),
+					      Qt::ElideRight,
+					      m_ui.value->width()));
+      m_ui.value->setToolTip(cookie.at(0).value());
       m_ui.value->setCursorPosition(0);
     }
 }
@@ -720,7 +1000,10 @@ void dooble_cookies_window::slot_purge_domains_timer_timeout(void)
     {
       auto item = m_ui.tree->topLevelItem(i);
 
-      if(!(!item || item->checkState(0) == Qt::Checked))
+      if(!item)
+	continue;
+
+      if(item->checkState(0) == Qt::Unchecked)
 	list << item;
     }
